backend/requirements.txt
Django==5.0.8
djangorestframework==3.15.2
django-cors-headers==4.4.0
channels==4.1.0
psycopg2-binary==2.9.9

faster-whisper==1.0.3
sentence-transformers==3.0.1
python-dotenv==1.0.1
Pillow==10.4.0
numpy==1.26.4
openai>=1.30.0
daphne>=4.1.0
